Added better url routes - slug instead of id and zasady-uczestnictwa/slug
