+++ src/app/szlak-partyzancki/rajdy/_models/rallyData.ts
@@ -1,5 +1,6 @@
 export interface RallyData {
   id: number;
+  slug: string;
   title: string;
   date: string;
   invite: string;
