+++ src/app/szlak-partyzancki/rajdy/data/rallyPageData.ts
@@ -3,6 +3,7 @@ import { Rally } from "../_components/RallyList";
 export const allRallies: Rally[] = [
   {
     id: 1,
+    slug: "rajd-kaluzowka-2018",
     title: "Rajd rowerowy na Kałużówkę 2018",
     date: "5 sierpnia, 2018",
     description: "W roku stulecia odzyskania niepodległości Polski serdecznie zapraszamy uczniów szkół podstawowych i klas gimnazjalnych na XIII Rajd Gwiaździsty „Kałużówka 2018”, z okazji Święta Pułkowego, o Przechodni Proporzec 5. Pułku Strzelców Konnych Armii Krajowej. Hasłem i motywem przewodnim podczas tegorocznej wędrówki jest „Pokolenie Niepodległej”.",
@@ -10,6 +11,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 2,
+    slug: "rajd-szlak-partyzancki-ponadgimnazjalne",
     title: "Rajd Szlakiem Partyzanckim dla uczniów szkół ponadgimnazjalnych",
     date: "11 października, 2017",
     description: "Katolickie Centrum Edukacji KANA w Dębicy, Stowarzyszenie im. Edmunda Bojanowskiego „Dobroć”, przy wsparciu finansowym Ministerstwa Obrony Narodowej oraz przy wsparciu organizacyjnym Światowego Związku Żołnierzy AK i Powiatu Dębickiego zapraszają młodzież szkół ponadgimnazjalnych na Rajd Szlakiem Partyzanckim II",
@@ -17,6 +19,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 3,
+    slug: "rajd-kaluzowka-2017",
     title: "Rajd rowerowy na Kałużówkę",
     date: "5 sierpnia, 2017",
     description: "W 73. rocznicę pacyfikacji Gumnisk i Braciejowej, 12 sierpnia 2017 roku, serdecznie zapraszamy wszystkich miłośników jazdy na rowerze na rajd rowerowy z Dębicy na Kałużówkę. Rajd wyruszy o godzinie 10:00 z parkingu przy Katolickim Centrum Edukacji KANA w Dębicy (ulica Chopina 2). Organizatorami rajdu są: Miasto Dębica, Katolickie Centrum Edukacji KANA, Światowy",
@@ -24,6 +27,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 4,
+    slug: "rajd-gwiazdzisty-2017",
     title: "12 czerwca zapraszamy na XII Rajd Gwiaździsty „KAŁUŻÓWKA 2017”",
     date: "23 maja, 2017",
     description: "Światowy Związek Żołnierzy Armii Krajowej Środowisko 5. Pułku Strzelców Konnych AK w Dębicy, Parafia pw. Wniebowzięcia Najświętszej Marii Panny w Gumniskach Zespół Szkół w Gumniskach zapraszają na XII Rajd Gwiaździsty „KAŁUŻÓWKA 2017” o Przechodni Proporzec 5. Pułku Strzelców Konnych AK, Szlakiem Partyzanckim.",
@@ -31,6 +35,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 5,
+    slug: "rajd-rozwiazanie-ak-2017",
     title: "IV rajd na rozwiązanie Armii Krajowej",
     date: "29 stycznia, 2017",
     description: "Kilkanaście osób, w tym żołnierze Narodowych Sił Zbrojnych, członkowie Światowego Związku Żołnierzy AK, strzelcy i osoby zainteresowane historią polskiego podziemia, wzięło udział, 22 stycznia 2017 roku, w czwartym rajdzie pieszym zorganizowanym w rocznicę rozwiązania Armii Krajowej. Rajd był także okazją do uczczenia pamięci powstańców styczniowych, w 154.",
@@ -38,6 +43,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 6,
+    slug: "rowerowy-rajd-inauguracja-2016",
     title: "Rowerowy rajd Szlakiem Partyzanckim – inauguracja trasy!",
     date: "17 września, 2016",
     description: "Katolickie Centrum Edukacji KANA w Dębicy – realizator projektu odnowienia Szlaku Partyzanckiego; Światowy Związek Żołnierzy Armii Krajowej – opiekun merytoryczny szlaku oraz Miasto Dębica, zapraszają na Rajd Rowerowy Szlakiem Partyzanckim II Zgrupowania Armii Krajowej Obwodu Dębica. Wydarzenie będzie jednocześnie inauguracją rowerowego",
@@ -45,6 +51,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 7,
+    slug: "rajd-gwiazdzisty-start-2016",
     title: "Informacja o miejscach wymarszu patroli na Rajd Gwiaździsty 2016",
     date: "8 czerwca, 2016",
     description: "Informujemy o miejscach startu poszczególnych patroli na XI Rajd Gwiaździsty, który odbędzie się w piątek 10 czerwca 2016 roku. Rajd tradycyjnie odbywał się będzie w ramach obchodów święta 5. Pułku Strzelców Konnych.",
@@ -52,6 +59,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 8,
+    slug: "rajd-gwiazdzisty-2016",
     title: "XI Rajd Gwiaździsty „KAŁUŻÓWKA 2016” – zapowiedź",
     date: "26 maja, 2016",
     description: "Światowy Związek Żołnierzy Armii Krajowej Środowisko 5. Pułku Strzelców Konnych AK w Dębicy, Parafia pw. Wniebowzięcia Najświętszej Marii Panny w Gumniskach, Zespół Szkół w Gumniskach oraz Katolickie Centrum Edukacji KANA zapraszają na XI Rajd Gwiaździsty „KAŁUŻÓWKA 2016” o Przechodni Proporzec 5. Pułku Strzelców Konnych „Kałużówka 2016” Szlakiem",
@@ -59,6 +67,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 9,
+    slug: "rajd-inauguracyjny-2016",
     title: "Zaproszenie na Rajd Inauguracyjny",
     date: "26 maja, 2016",
     description: "Katolickie Centrum Edukacji KANA w Dębicy, Światowy Związek Żołnierzy Armii Krajowej Środowisko 5. Pułku Strzelców Konnych AK w Dębicy zapraszają patrole harcerskie, partole Związku Strzeleckiego oraz młodzież szkół ponadgimnazjalnych z powiatu dębickiego na Rajd Inauguracyjny odnowionego Szlaku Partyzanckiego II Zgrupowania Armii Krajowej Obwodu Dębica.",
@@ -66,6 +75,7 @@ export const allRallies: Rally[] = [
   },
   {
     id: 10,
+    slug: "rajd-rozwiazanie-ak-2016",
     title: "III Rajd na rozwiązanie Armii Krajowej",
     date: "22 stycznia, 2016",
     description: "17 stycznia 2016 roku odbył się trzeci rajd w rocznicę rozwiązania Armii Krajowej. W rajdzie wzięli udział członkowie Związku Strzeleckiego z Dębicy, a także sympatycy ideałów Armii Krajowej i członkowie Światowego Związku Żołnierzy Armii Krajowej w Dębicy. Uczestnicy przeszli trasę 10 km, fragmentem Szlaku Partyzanckiego II Zgrupowania Armii Krajowej Obwodu",
